Create 'apply-processor' processor branch.
In this branch, I will factor-out an "apply" processor from the merge command
implementation. This will help us using it in different ways, for example, for
xpacth feature, explained in [1].

Access more information about it in the same thread on the dev@ mailing
list: [1]

This is a short living branch, existing to ensure that all tests are passing,
and splitting a huge diff, we will have for few commits.

[1] https://lists.apache.org/thread/cjldotpstz60fr0v330zrkprd9zjw0q5

git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/apply-processor@1922042 13f79535-47bb-0310-9956-ffa450edef68