Follow-up to r1920955: Fix broken autoconf out-of-tree builds.
The changes in that commit removed the configuration of svneditor script. This
followed that we are no longer having svneditor script in the build-dir.

Previously, the path to svneditor_script has been determined using process'
current directory on Unix, while the Windows code is already determining it by
joining the current script path and relative path to this script, taking this
file from the source-dir.

For instance, it has been working on in-tree build, because the build-dir and
the source-dir have actually been the same, so we haven't reproduced this
problem before.

This commit will follow the function, determining the absolute path to the
svneditor script, using the same approach on Unix and Windows platforms.

This has been reported and discussed on dev@. The discussion is archived and
can be accessed at: [1]

* subversion/tests/cmdline/svntest/main.py
  (svneditor_script (non-windows)): Determine the script's root directory
   from `sys.path[0]`.

Found by: danielsh
Reviewed by: jun66j5
             dsahlberg
             hartmannathan
Tested by: jun66j5

[1] https://lists.apache.org/thread/q6bo8hjks16bn2ypnk9x1ys7km5spyqy


git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@1922113 13f79535-47bb-0310-9956-ffa450edef68