On the 'apply-processor' branch: Fix replace-style merge capability for a single file.
This fixes test cmdline/merge_tests.py#52 (replace-style merge capability
for a single file) and implements TODO_REPLACE_SINGLE_FILE todo.

For this we'll be opening a directory with NULL, passed to the relpath
argument. This tells the apply processor to consider this directory as
"fake", which is opened, but doesn't exist either in the working copy or
the repository. In do_file_merge() function, we will use processor->dir_opened()
function to open the parent directory. For instance, before we were using the
open_dir_for_replace_single_file() function, that creates a new fake baton.
The updated implementation of dir_opened will check relpath for NULL, and
do easy-out if it is NULL. However, new_dir_baton will be initialized with
the a fake dir baton.

This replaces the open_dir_for_replace_single_file() function, so I'm going
to remove this function, because it is no longer referenced.

* subversion/libsvn_client/merge.c
  (do_file_merge): Remove TODO_REPLACE_SINGLE_FILE, use processor->dir_opened()
   instead of open_dir_for_replace_single_file() to open the dir baton, little
   rearrangement of local variables.
* subversion/libsvn_client/merge_processor.c
  (merge_dir_opened): Easy-out if relapth is NULL; Only creates a "fake" dir
   baton.
  (open_dir_for_replace_single_file): Remove function.

### I am not sure in that. This works, but probably it is not the best
### solution. If anyone want to express their opinion about this change,
### feel free to do so.

### Is NULL `relpath` conceptually correct? Does it follow
### svn_diff_tree_processor_t's conventions?

git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/apply-processor@1922134 13f79535-47bb-0310-9956-ffa450edef68