On the 'apply-processor' branch: Notice the paths in merged_abspaths on each addition.
During the check of differences between the 'apply-processor' branch and the
trunk, I noticed that it was accidently changed somehow. Reverting it back
now.

This is a follow-up to r1922090 (add callbacks for notifying merge.c
from merge_processor.c), where the problem has been introduced into the
codebase.

In my testings, I got the test-suite running on all-green, will check the
GitHub Actions status as soon as it would be ready!

* subversion/libsvn_client/merge.c
  (apply_processor_updated_path:svn_wc_notify_update_add): Store the path
   in merge_b->merged_abspaths.

git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/apply-processor@1922202 13f79535-47bb-0310-9956-ffa450edef68