On Windows, use the WC_NO_BEST_FIT_CHARS [1] flag when converting individual command-line arguments to the ANSI code page.
If a best-fit mapping occurs, return an error instead of silently accepting
the results of the mapping.  This avoids unintended input changes (e.g., a
fullwidth hyphen-minus being converted to a regular hyphen).

This is an interim fix. A long-term solution is to convert arguments to UTF-8
and work with UTF-8 directly, but until then failing fast should be better
than changing the user input.

* subversion/libsvn_subr/cmdline.c
  (svn_cmdline__win32_get_cstring_argv): Specify WC_NO_BEST_FIT_CHARS during
   conversion and fail if best-fit mapping occurs.

* subversion/tests/cmdline/basic_tests.py
  (argv_with_best_fit_chars): Update test expectations, and expect an error.
   Check only the error prefix, since the exact message and its OS portion
   may vary due to localization.

[1]: https://learn.microsoft.com/en-us/windows/win32/api/stringapiset/nf-stringapiset-widechartomultibyte


git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@1927953 13f79535-47bb-0310-9956-ffa450edef68