ra_serf: Implement and start using a default readline handler for custom bucket types.
Before this change, we were making an assumption that the calling site
is not going to call bucket->readline() on our custom bucket types.
This is not something a bucket type has control over, and the assumption
may not hold in the future.

To solve this issue, we provide a default bucket readline implementation,
modeled over the current implementation of serf_default_readline() in
serf trunk.

* subversion/libsvn_ra_serf/ra_serf.h
  (svn_ra_serf__default_readline): New function.

* subversion/libsvn_ra_serf/util.c
  (svn_ra_serf__default_readline): Implement based on the current
   implementation of serf_default_readline() in serf trunk.

* subversion/libsvn_ra_serf/eagain_bucket.c
* subversion/libsvn_ra_serf/sb_bucket.c:
  Use svn_ra_serf__default_readline() as the readline handler.


git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@1930807 13f79535-47bb-0310-9956-ffa450edef68