Add OpenSSL backend for checksum computation.
After this change, every component that relies on checksuming would improve in
performance by a lot in thanks to OpenSSL optimised hash algorithms. My
benchmarks showed the following improvements:

- The benchmarking test was showing 0.16 GB/s raw speed with APR backend and
  1.38 seconds GB/s with OpenSSL. All of those are SHA1. This is 8.62 times jump!

- Status of a working copy with a single file (with size of 370 MB) simply
  touched was taking 1.94 seconds of real time to execute with APR backend and
  0.30 seconds with OpenSSL (6.46 times improvement).

- The same test with a file of around 3 GB had a jump from 15.73 seconds to
  2.23 seconds (7.05 times improvement).

- Checkout of the whole repository with those two files over file:// protocol
  took 40 seconds with APR and 20 seconds with OpenSSL (2 times improvement).

Currenly, only CMake build system implements support for this feature. As far
as I checked, it should not at least break other build systems, but I didn't
implement support for them either.

This implementation relies on deprecated API of OpenSSL which (I think) is
objectively better than the approach they currenly consider recommended (which
is to use EVP interface).

1) it's just simpler
2) doesn't perform vtable calls
3) and OpenSSL uses it anyway in EVP

(personally, i don't get the reason why it was deprecated in the first place)

This is a completelly optional feature which is disabled by default. If one
doesn't want to use it, we wont force you to. There are many disadvantages of
loading such heavy libraries when we don't need all the stuff it gives us. This
optimisation comes with its own cost. But, our client already loads OpenSSL
indirectly through libserf so this is not an issue in some cases. Let's just
try it. Why not?

* CMakeLists.txt
  (options): Add SVN_CHECKSUM_USE_OPENSSL.
  (SVN_CHECKSUM_USE_OPENSSL): Check for option to decide whether to look for
   the OpenSSL library and setup definitions or not.
  (configsummary): Log value of SVN_CHECKSUM_USE_OPENSSL.

* build.conf
  (libsvn_subr): Add dependency on openssl.

* subversion/libsvn_subr/checksum_apr.c
  (): Put the whole file into an ifndef condition depending on
   SVN_CHECKSUM_USE_OPENSSL.

* subversion/libsvn_subr/checksum_openssl.c
  (svn_checksum__md5,
   svn_checksum__md5_ctx_t,
   svn_checksum__md5_ctx_create,
   svn_checksum__md5_ctx_reset,
   svn_checksum__md5_ctx_update,
   svn_checksum__md5_ctx_final,
   svn_checksum__sha1,
   svn_checksum__sha1_ctx_t,
   svn_checksum__sha1_ctx_create,
   svn_checksum__sha1_ctx_reset,
   svn_checksum__sha1_ctx_update,
   svn_checksum__sha1_ctx_final): Implement through OpenSSL (under an ifdef).


git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@1930949 13f79535-47bb-0310-9956-ffa450edef68